Update Cases with Portal.adoc
Sorry my stars (quite literally) are probably not aligned very well here but it's important that end users understand the limits of the portal option before spending time on it to find that it needs significant development. Not everybody is a coder and it's important that end users are not disappointed by it. If the additions appear to critical, it may be time to consider totally pulling the portal functionality as I understand that it cannot be a priority for the development team due to the backlogs and other issues contrasted with the small size of the development team. I have included an invite for help with the code within this proposed update in the hope that someone else can take charge of this that is a coder as a volunteer. I only wish that I saw this over a week ago as I realise that I am going to need to entirely replace Suite CRM with an alternative service.